move grammar and eval semantics into new files
