+++ SiliconGrammar.ts
@@ -1,0 +1,4 @@
+import * as ohm from 'ohm-js'
+const grammarSource = Bun.file('./src/silicon-simple.ohm')
+const siliconGrammar = ohm.grammar(await grammarSource.text());
+export default siliconGrammar
